update powershell script to configure workspace
Azure PowerShell team is removing Switch-AzureMode, which will be a breaking change to create a solid foundation for the future of Azure PowerShell. The changes are as follows and are targeted for release on Oct. 8th of 2015.

(as mentioned in https://github.com/Azure/azure-powershell/wiki/Deprecation-of-Switch-AzureMode-in-Azure-PowerShell)

Example script was modified to use new AzureRm* commands